move enable trail action to account resource
