+++ tests/data/placebo/test_cloudtrail_enable/cloudtrail.CreateTrail_1.json
@@ -1,0 +1,22 @@
+{
+    "status_code": 200, 
+    "data": {
+        "IncludeGlobalServiceEvents": true, 
+        "Name": "nosetest", 
+        "ResponseMetadata": {
+            "RetryAttempts": 0, 
+            "HTTPStatusCode": 200, 
+            "RequestId": "9830fd0d-ed81-11e6-9205-e16fedb50da3", 
+            "HTTPHeaders": {
+                "x-amzn-requestid": "9830fd0d-ed81-11e6-9205-e16fedb50da3", 
+                "date": "Tue, 07 Feb 2017 22:05:53 GMT", 
+                "content-length": "216", 
+                "content-type": "application/x-amz-json-1.1"
+            }
+        }, 
+        "TrailARN": "arn:aws:cloudtrail:us-east-1:644160558196:trail/nosetest", 
+        "LogFileValidationEnabled": false, 
+        "IsMultiRegionTrail": false, 
+        "S3BucketName": "nosetest-bucket"
+    }
+}
